+++ conftest.py
@@ -3,6 +3,7 @@ import hypothesis
 import pytest
 from astropy.utils.data import check_download_cache
 from astropy.config import paths
+import pint.utils
 
 # This setup is drawn from Astropy and might not be entirely relevant to us;
 # in particular we don't have a cron run for slow tests.
@@ -42,3 +43,7 @@ def temp_cache(tmpdir):
     with paths.set_temp_cache(tmpdir):
         yield None
         check_download_cache()
+
+
+# Refuse to run test suite if precision not available
+pint.utils.require_longdouble_precision()